requirements.txt
email-validator>=2.3.0
flask>=3.1.2
flask-sqlalchemy>=3.1.1
gtts>=2.5.4
gunicorn>=23.0.0
psycopg2-binary>=2.9.10
requests>=2.32.5
sqlalchemy>=2.0.43
python-dotenv>=1.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
# stability-sdk>=0.2.0
google-cloud-texttospeech>=2.14.0
protobuf>=4.25.0
huggingface_hub>=0.34.0
moviepy>=1.0.3